DnfContext: improve error-reporting for modular functions
As a library, we need to be able to report errors directly to the
application and not log it ourselves. Applications should take care of
displaying the errors to users.

Additionally, let's not log critical errors for things which are due to
user-input, such as a non-existent module name. In rpm-ostree for
example, we run enable `G_DEBUG=fatal-warnings` in some test scenarios
to ensure that we don't trigger any warnings. This then causes
rpm-ostree to crash if doing `rpm-ostree module install badname`.

This patch changes things so that instead of logging errors, we
concatenate them together and return them through the GError API.